+++ src/OmniSharp.Roslyn/Extensions/SymbolExtensions.cs
@@ -1,6 +1,9 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
+using System.IO;
+using System.Linq;
+using System.Text;
 using Microsoft.CodeAnalysis;
 using OmniSharp.Models.V2;
 using OmniSharp.Roslyn.Utilities;
@@ -247,5 +250,52 @@ namespace OmniSharp.Extensions
 
             return (INamedTypeSymbol)topLevelNamedType;
         }
+
+        public static string GetSymbolName(this ISymbol symbol)
+        {
+            var topLevelSymbol = symbol.GetTopLevelContainingNamedType();
+            return GetTypeDisplayString(topLevelSymbol);
+        }
+
+        private static string GetTypeDisplayString(INamedTypeSymbol symbol)
+        {
+            if (symbol.SpecialType != SpecialType.None)
+            {
+                var specialType = symbol.SpecialType;
+                var name = Enum.GetName(typeof(SpecialType), symbol.SpecialType).Replace("_", ".");
+                return name;
+            }
+
+            if (symbol.IsGenericType)
+            {
+                symbol = symbol.ConstructUnboundGenericType();
+            }
+
+            if (symbol.IsUnboundGenericType)
+            {
+                // TODO: Is this the best to get the fully metadata name?
+                var parts = symbol.ToDisplayParts();
+                var filteredParts = parts.Where(x => x.Kind != SymbolDisplayPartKind.Punctuation).ToArray();
+                var typeName = new StringBuilder();
+                foreach (var part in filteredParts.Take(filteredParts.Length - 1))
+                {
+                    typeName.Append(part.Symbol.Name);
+                    typeName.Append(".");
+                }
+                typeName.Append(symbol.MetadataName);
+
+                return typeName.ToString();
+            }
+
+            return symbol.ToDisplayString();
+        }
+
+        internal static string GetFilePathForExternalSymbol(this ISymbol symbol, Project project)
+        {
+            var topLevelSymbol = symbol.GetTopLevelContainingNamedType();
+            return $"$metadata$/Project/{Folderize(project.Name)}/Assembly/{Folderize(topLevelSymbol.ContainingAssembly.Name)}/Symbol/{Folderize(GetTypeDisplayString(topLevelSymbol))}.cs".Replace(Path.AltDirectorySeparatorChar, Path.DirectorySeparatorChar);
+        }
+
+        private static string Folderize(string path) => string.Join("/", path.Split('.'));
     }
 }
